Tests: * Update StaticPrimesTest - move to Primes package * Update PrimeValidation Tests - move to Primes package * Update BytePrimeCacheTest - optimize imports
