+++ numbers/src/test/java/mathtools/numbers/primes/PrimeValidationIsPrimeTest.java
@@ -1,16 +1,13 @@
-package mathtools.numbers.primes.validate;
+package mathtools.numbers.primes;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static mathtools.numbers.primes.validate.PrimeValidation.isPrime;
+import static mathtools.numbers.primes.PrimeValidation.isPrime;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import mathtools.numbers.primes.BytePrimeCache;
-import mathtools.numbers.primes.PrimeCacheInterface;
 
 /** Testing [PrimeValidation] IsPrime Method
  * @author DK96-OS : 2022 */
